add index.ts barrel files for @modelcontextprotocol/sdk/server, @modelcontextprotocol/sdk/client and @modelcontextprotocol/sdk/shared
