+++ src/server/auth/index.ts
@@ -1,0 +1,21 @@
+// Auth exports - root auth level
+export * from './clients.js';
+export * from './errors.js';
+export * from './provider.js';
+export * from './router.js';
+export * from './types.js';
+
+// Auth exports - handlers
+export * from './handlers/authorize.js';
+export * from './handlers/metadata.js';
+export * from './handlers/register.js';
+export * from './handlers/revoke.js';
+export * from './handlers/token.js';
+
+// Auth exports - middleware
+export * from './middleware/allowedMethods.js';
+export * from './middleware/bearerAuth.js';
+export * from './middleware/clientAuth.js';
+
+// Auth exports - providers
+export * from './providers/proxyProvider.js';
