+++ src/server/server.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable no-constant-binary-expression */
 /* eslint-disable @typescript-eslint/no-unused-expressions */
-import { Server } from "./index.js";
+import { Server } from "./server.js";
 import { z } from "zod";
 import {
   RequestSchema,
