+++ src/shared/index.ts
@@ -1,0 +1,7 @@
+export * from './auth-utils.js';
+export * from './auth.js';
+export * from './metadataUtils.js';
+export * from './protocol.js';
+export * from './stdio.js';
+export * from './transport.js';
+export * from './uriTemplate.js';
